requirements.txt
Flask==2.3.2
gspread==5.12.0
google-auth==2.28.1
